aooga

+++ helpers.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-double euclid_distance(tuple<int,int,int,int> p1, tuple<int,int,int,int> p2) {
+double euclid_distance(tuple<double,double,double,double> p1, tuple<double,double,double,double> p2) {
     double dist_sum = pow((get<0>(p1) - get<0>(p2)),2) + 
                       pow((get<1>(p1) - get<1>(p2)),2) +
                       pow((get<2>(p1) - get<2>(p2)),2) +
